% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer_example.R
\docType{data}
\name{buffer_example}
\alias{buffer_example}
\title{Example bufferstrips}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 84 columns.
}
\usage{
buffer_example
}
\description{
This is a subset of the \link{bufferstrips} datatset used in the walkthroughs.
}
\keyword{datasets}
