% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buffer_example_long.R
\docType{data}
\name{buffer_example_long}
\alias{buffer_example_long}
\title{Long-form example}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 97 rows and 16 columns.
}
\usage{
buffer_example_long
}
\description{
From \link{bufferstrips} dataset.
}
\keyword{datasets}
