% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bufferstrips.R
\docType{data}
\name{bufferstrips}
\alias{bufferstrips}
\title{Bufferstrips dataset}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 1072 rows and 84 columns.
}
\usage{
bufferstrips
}
\description{
todo: source
}
\details{
See data-raw/example_datasets.R for details about how this dataset is cleaned for
import.
}
\keyword{datasets}
