% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Check_condensed.R
\name{check_condensed}
\alias{check_condensed}
\title{Check dataframe for multiple possible separators}
\usage{
check_condensed(data, value.sep = ";")
}
\arguments{
\item{data}{A data frame}

\item{value.sep}{An expected column separator}
}
\description{
Check dataframe for multiple possible separators
}
\examples{
data(bufferstrips)
check_condensed(bufferstrips, value.sep = ';')
}
