% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_condensed.R
\name{condensed_to_long}
\alias{condensed_to_long}
\title{Condensed to long}
\usage{
condensed_to_long(data, columns, value_sep = ";")
}
\arguments{
\item{data}{A dataframe containing rows of systematic review data in condensed
format.}

\item{columns}{A set of columns containing coded data from within a dataframe.}

\item{value_sep}{A character used to separate values within the data. The
default is set to ';'.}
}
\value{
A dataframe in 'long' format.
}
\description{
Convert from condensed format to long format based on the presence
of a given value separator within a given set of columns.
}
\examples{
fpath <- system.file("extdata", "data.csv", package="sysrevdata")
data <- read.csv(fpath)
columns <- c('region', 'hazard', 'response', 'sdgs', 'actor')
long_data <- condensed_to_long(data, columns, value_sep = 'xxx');
}
