% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_condensed.R
\name{condensed_to_wide}
\alias{condensed_to_wide}
\title{Condensed to wide}
\usage{
condensed_to_wide(data, columns, name_sep = "_", value_sep = ";")
}
\arguments{
\item{data}{A dataframe containing rows of systematic review data in condensed
format.}

\item{columns}{A set of columns containing coded data from within a dataframe.}

\item{name_sep}{A character used to separate terms in the newly generated column
names within the data. The default is set to '_' (i.e. snake_case).}

\item{value_sep}{A character used to separate values within the data. The
default is set to ';'.}
}
\value{
A dataframe in 'wide' format.
}
\description{
Convert from condensed format to wide format based on the presence
of a given value separator within a given set of columns.
}
\examples{
columns <- c('region', 'hazard', 'response', 'sdgs', 'actor')
wide_data <- condensed_to_wide(data, columns, value_sep = 'xxx');
}
