% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.R
\docType{data}
\name{hazard}
\alias{hazard}
\title{Hazard dataset}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1682 rows and 8 columns.
}
\usage{
hazard
}
\description{
todo: source
}
\details{
See data-raw/example_datasets.R for details about how this dataset is cleaned for
import.
}
\keyword{datasets}
