% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{is_condensed}
\alias{is_condensed}
\title{Check if database is in 'condensed' format}
\usage{
is_condensed(columns, value_sep = ";")
}
\arguments{
\item{columns}{A set of columns containing coded data from within a dataframe.}

\item{value_sep}{A character used to separate values within the data.}
}
\value{
Logical argument, TRUE or FALSE
}
\description{
Checks if separator is present in given coding columns and returns
TRUE or FALSE.
}
\examples{
is_condensed(columns, value_sep = ';');
}
