% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{is_long}
\alias{is_long}
\title{Check if database is in 'long' format}
\usage{
is_long(key_variable, columns, value_sep = ";")
}
\arguments{
\item{key_variable}{A variable used to identify unique articles/studies.}

\item{columns}{A set of columns containing coded data from within a dataframe}

\item{value_sep}{A character used to separate values within the data.}
}
\value{
Logical argument, TRUE or FALSE
}
\description{
Checks whether key.variable is duplicated AND there are no separated
values in coding variables and returns TRUE of FALSE.
}
\examples{
is_long(columns, value_sep = ';');
}
