% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair_and_replace.R
\name{replace_sep}
\alias{replace_sep}
\title{Replace a given separator across dataframe}
\usage{
replace_sep(data, old_sep, new_sep = "; ")
}
\arguments{
\item{data}{A dataframe containing rows of systematic review data.}

\item{old_sep}{An existing character used to separate values within the
data. The default is set to ';'.}

\item{new_sep}{A replacement character used to separate values within the
data. The default is set to ';'.}
}
\description{
Takes a given separator and replaces it with an alternative
across the dataframe.
}
\examples{
\dontrun{
data <- data \%>\%
    replace_sep(old_sep = 'xxx', new_sep = '; ');
    }
}
