% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair_and_replace.R
\name{replace_string}
\alias{replace_string}
\title{Replace text string across database}
\usage{
replace_string(data, replace, with)
}
\arguments{
\item{data}{A dataframe containing rows of systematic review data.}

\item{replace}{A string to be replaced across the dataframe.}

\item{with}{A string used to replace the above string.}
}
\value{
A dataframe with the string replaced.
}
\description{
Replace a given string.
}
\examples{
data <- replace_string(data, replace = 'NULL', with = '');
}
