% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair_and_replace.R
\name{trim_spaces}
\alias{trim_spaces}
\title{Remove whitespace}
\usage{
trim_spaces(data, value_sep = ";")
}
\arguments{
\item{data}{A dataframe containing rows of systematic review data.}

\item{value_sep}{A character used to separate values within the data. The
default is set to ';'.}
}
\value{
Dataframe with whitespace removed.
}
\description{
Remove 'unnecessary' whitespace (leading spaces and trailing
spaces for each cell value and either side of a given separator).
}
\examples{
data <- trim_spaces(data);
}
