% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_wide.R
\name{wide_to_condensed}
\alias{wide_to_condensed}
\title{Wide to condensed}
\usage{
wide_to_condensed(
  data,
  columns,
  value_sep = ";",
  name_sep = "_",
  readable = TRUE
)
}
\arguments{
\item{data}{A dataframe containing rows of systematic review data in wide
format.}

\item{columns}{A set of columns containing coded data from within a dataframe.}

\item{value_sep}{A character used to separate values within the data. The
default is set to ';'.}

\item{name_sep}{A character used to separate terms in the newly generated column
names within the data. The default is set to '_' (i.e. snake_case).}

\item{readable}{Logical argument (TRUE or FALSE) specifying whether the value separator
should be followed by a space to improve readability (e.g. '; ' instead of ';').
The default is set to 'readable = TRUE'.}
}
\value{
A dataframe in 'condensed' format.
}
\description{
Convert from wide format to condensed format based on the prefix for
each set of linked values across multiple columns.
}
\examples{
wide <- wide_to_condensed(data, columns, readable = TRUE)
wide;
}
