% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_wide.R
\name{wide_to_long}
\alias{wide_to_long}
\title{Wide to long}
\usage{
wide_to_long(data, column_stem, name_sep = "_")
}
\arguments{
\item{data}{A dataframe containing rows of systematic review data in wide
format.}

\item{column_stem}{A set of column descriptors signifying the groups of columns to condense
based on a common stem (e.g. 'hazard' represents 'hazard_low', 'hazard_moderate', and
'hazard_high').}

\item{value_sep}{A character used to separate values within the data. The
default is set to ';'.}
}
\value{
A dataframe in 'long' format.
}
\description{
Convert from wide format to long format based on the prefix for
each set of linked values across multiple columns.
}
\examples{
\dontrun{
long <- wide_to_long(data, column_stem)
long;
}
}
