% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_condensed}
\alias{check_condensed}
\title{Check dataframe for possible value separators}
\usage{
check_condensed(data, value_sep = ";")
}
\arguments{
\item{data}{A dataframe containing rows of systematic review data}

\item{value_sep}{A character used to separate values within the data. The
default is set to ';'.}
}
\value{
A dataframe listing all columns in the input data, with a column
showing whether the given (or default) value separator was detected in each,
and a column showing possible candidate separator (ie. non-alphanumeric)
characters found within each column of the input data.
}
\description{
Check dataframe for possible value separators
}
\examples{
check_condensed(data, value_sep = ';');
}
