% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_wide.R
\name{wide_to_condensed}
\alias{wide_to_condensed}
\title{Wide to condensed}
\usage{
wide_to_condensed(
  data,
  column_stem,
  value_sep = ";",
  name_sep = "_",
  readable = TRUE
)
}
\arguments{
\item{data}{A dataframe containing rows of systematic review data in wide
format.}

\item{column_stem}{A set of column descriptors signifying the groups of columns to condense
based on a common stem (e.g. 'hazard' represents 'hazard_low', 'hazard_moderate', and
'hazard_high').}

\item{value_sep}{A character used to separate values within the data. The
default is set to ';'.}

\item{name_sep}{A character used to separate terms in the newly generated column
names within the data. The default is set to '_' (i.e. snake_case).}

\item{readable}{Logical argument (TRUE or FALSE) specifying whether the value separator
should be followed by a space to improve readability (e.g. '; ' instead of ';').
The default is set to 'readable = TRUE'.}
}
\value{
A dataframe in 'condensed' format.
}
\description{
Convert from wide format to condensed format based on the prefix for
each set of linked values across multiple columns.
}
\examples{
column_stem <- 'hazard'
condensed <- wide_to_condensed(data, column_stem, readable = TRUE)
condensed;
}
