% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_wide.R
\name{wide_to_long}
\alias{wide_to_long}
\title{Wide to long}
\usage{
wide_to_long(data, columns, name_sep = "_")
}
\arguments{
\item{data}{A dataframe containing rows of systematic review data in wide
format.}

\item{columns}{A set of columns containing coded data from within a dataframe.}

\item{value_sep}{A character used to separate values within the data. The
default is set to ';'.}
}
\value{
A dataframe in 'long' format.
}
\description{
Convert from wide format to long format based on the prefix for
each set of linked values across multiple columns.
}
\examples{
long <- wide_to_long(data, columns)
long;
}
